fix(seed):moved seed to script folder
