+++ script/seed.js
@@ -1,51 +1,43 @@
-/**
- * Welcome to the seed file! This seed file uses a newer language feature called...
- *
- *                  -=-= ASYNC...AWAIT -=-=
- *
- * Async-await is a joy to use! Read more about it in the MDN docs:
- *
- * https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Statements/async_function
- *
- * Now that you've got the main idea, check it out in practice below!
- */
-const db = require('../server/db')
-const {User} = require('../server/db/models')
+const db = require('./server/db/db');
+const User = require('./server/db/models/user');
+const Product = require('./server/db/models/product');
+const Faker = require('faker');
 
-async function seed () {
-  await db.sync({force: true})
-  console.log('db synced!')
-  // Whoa! Because we `await` the promise that db.sync returns, the next line will not be
-  // executed until that promise resolves!
-
-  const users = await Promise.all([
-    User.create({email: 'cody@email.com', password: '123'}),
-    User.create({email: 'murphy@email.com', password: '123'})
-  ])
-  // Wowzers! We can even `await` on the right-hand side of the assignment operator
-  // and store the result that the promise resolves to in a variable! This is nice!
-  console.log(`seeded ${users.length} users`)
-  console.log(`seeded successfully`)
+function users() {
+  Faker.seed(123);
+  const x = [];
+  for (let i = 0; i < 5; i++) {
+    x.push(User.create({
+      name: Faker.name.findName(),
+      email: Faker.internet.email(),
+      googleId: Faker.internet.userName(),
+      salt: Faker.image.avatar(),
+      password: Faker.lorem.word(1),
+    }));
+  }
+  return x;
 }
 
-// Execute the `seed` function
-// `Async` functions always return a promise, so we can use `catch` to handle any errors
-// that might occur inside of `seed`
-seed()
-  .catch(err => {
-    console.error(err.message)
-    console.error(err.stack)
-    process.exitCode = 1
-  })
-  .then(() => {
-    console.log('closing db connection')
-    db.close()
-    console.log('db connection closed')
-  })
+function product() {
+  Faker.seed(123);
+  const products = [];
+  for (let i = 0; i < 25; i++) {
+    products.push(Product.create({
+      name: Faker.commerce.productName(),
+      price: Faker.commerce.price(),
+      description: Faker.lorem.paragraph(3),
+      inventory: Math.floor(Math.random() * 4) + 1,
+    }));
+  }
+  return products;
+}
 
-/*
- * note: everything outside of the async function is totally synchronous
- * The console.log below will occur before any of the logs that occur inside
- * of the async function
- */
-console.log('seeding...')
+db
+  .sync({ force: true })
+  .then(() => Promise.all(users()))
+  .then(() => Promise.all(product()))
+  .then(() => {
+    console.log('Seeding successful');
+    db.close();
+    return null;
+  });
